- make environment non abstract and remove implementations from pkm-main - move environment introspection into pkm, remove dependencies - implement compatibility tag ordering logic - read and expose package metadata - the environment `installed_packages` can be pre-implemented as part of the api (since we now know how get it without   running the process and also have the ability to get the sitepackages path from the api) - virtual environment - remove

+++ pkm/api/packages/site_packages.py
@@ -15,22 +15,41 @@ if TYPE_CHECKING:
 
 
 class SitePackages:
-    def __init__(self, sites: Iterable[Path]):
-        self._sites = sites
+    def __init__(self, purelib: Path, platlib: Path, other_sites: Iterable[Path], is_read_only: bool):
+        self._purelib = purelib
+        self._platlib = platlib
+        self._other_sites = other_sites
+        self._is_read_only = is_read_only
+
+    @property
+    def purelib_path(self) -> Path:
+        return self._purelib
+
+    @property
+    def platlib_path(self) -> Path:
+        return self._platlib
 
     @cached_property
     def _name_to_packages(self) -> Dict[str, "InstalledPackage"]:
         result: Dict[str, InstalledPackage] = {}
-        for site in self._sites:
-            for file in site.iterdir():
-                if file.suffix == ".dist-info":
-                    metadata_file = file / "METADATA"
-                    if metadata_file.exists():
-                        metadata = PackageMetadata.load(metadata_file)
-                        user_request = _read_user_request(file, metadata)
-                        result[metadata.package_name] = InstalledPackage(file, metadata, user_request, self)
+        self._scan_packages(self._purelib, result, self._is_read_only)
+        self._scan_packages(self._platlib, result, self._is_read_only)
+        for other_site in self._other_sites:
+            self._scan_packages(other_site, result, True)
 
         return result
+
+    def _scan_packages(self, site: Path, result: Dict[str, "InstalledPackage"], readonly: bool):
+        if not site.exists():
+            return
+
+        for file in site.iterdir():
+            if file.suffix == ".dist-info":
+                metadata_file = file / "METADATA"
+                if metadata_file.exists():
+                    metadata = PackageMetadata.load(metadata_file)
+                    user_request = _read_user_request(file, metadata)
+                    result[metadata.package_name] = InstalledPackage(file, metadata, user_request, self, readonly)
 
     def installed_packages(self) -> Iterable["InstalledPackage"]:
         return self._name_to_packages.values()
@@ -61,7 +80,7 @@ class InstalledPackage(Package):
 
     def __init__(
             self, dist_info: Path, metadata: PackageMetadata, user_request: Optional[Dependency],
-            site: SitePackages):
+            site: SitePackages, readonly: bool):
 
         self._meta = metadata
         self._desc = PackageDescriptor(metadata.package_name, metadata.package_version)
@@ -69,6 +88,7 @@ class InstalledPackage(Package):
         self._dist_info = dist_info
         self._user_request = user_request
         self.site = site
+        self.readonly = readonly
 
     @property
     def descriptor(self) -> PackageDescriptor:
@@ -78,13 +98,21 @@ class InstalledPackage(Package):
     def user_request(self) -> Optional[Dependency]:
         return self._user_request
 
-    def unmark_user_requested(self):
+    def unmark_user_requested(self) -> bool:
+        if self.readonly:
+            return False
+
         (self._dist_info / "REQUESTED").unlink()
         self._user_request = None
+        return True
 
-    def mark_user_requested(self, request: Dependency):
+    def mark_user_requested(self, request: Dependency) -> bool:
+        if self.readonly:
+            return False
+
         (self._dist_info / "REQUESTED").write_text(str(request))
         self._user_request = request
+        return True
 
     def _all_dependencies(self, environment: "Environment") -> List[Dependency]:
         return self._meta.dependencies
@@ -95,14 +123,17 @@ class InstalledPackage(Package):
     def install_to(self, env: "Environment", user_request: Optional[Dependency] = None):
         raise NotImplemented()  # maybe re-mark user request?
 
-    def uninstall(self):
+    def uninstall(self) -> bool:
+        if self.readonly:
+            return False
+
         root = self._dist_info.parent
 
         parents: Set[Path] = set()
         record_file = self._dist_info / "RECORD"
 
         if not record_file.exists():
-            return  # unremovable package - should i notify it?
+            return False
 
         with record_file.open('r', newline='') as record_fd:
             for record in csv.reader(record_fd):
@@ -128,3 +159,5 @@ class InstalledPackage(Package):
                     path.rmdir()
 
         self.site.reload()
+
+        return True
